requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
